lichess_api_downloader/pyproject.toml
[project]
name = "lichess-downloader"
version = "0.1.0"
description = "Download Italian Game (C50-C59) games from Lichess for Neo4j loading"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "python-dotenv>=1.0",
]

[project.scripts]
lichess-downloader = "lichess_downloader.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
